requirements.txt
streamlit
python-dotenv
openai
streamlit_ace
